requirements.txt
fastapi
uvicorn
groq
numpy
pydantic
scikit-learn
pandas
tavily-python
python-dotenv